+++ src-tauri/src/puzzle.rs
@@ -107,7 +107,7 @@ pub struct PuzzleDatabaseInfo {
     title: String,
     description: String,
     puzzle_count: i32,
-    storage_size: u64,
+    storage_size: f64,
     path: String,
 }
 
@@ -121,7 +121,7 @@ pub async fn get_puzzle_db_info(file: PathBuf) -> Result<PuzzleDatabaseInfo, Err
 
     let puzzle_count = puzzles::table.count().get_result::<i64>(&mut db)? as i32;
 
-    let storage_size = path.metadata()?.len();
+    let storage_size = path.metadata()?.len() as f64;
     let filename = path.file_name().expect("get filename").to_string_lossy();
 
     Ok(PuzzleDatabaseInfo {
